Added Appbar with like and unlike navigation
