+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/petit/projects/react/chunknorrisassesment/src/index.js":"1","/home/petit/projects/react/chunknorrisassesment/src/App.js":"2"},{"size":428,"mtime":1627977527326,"results":"3","hashOfConfig":"4"},{"size":5307,"mtime":1629123027705,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"8"},"x0imxm",{"filePath":"9","messages":"10","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/petit/projects/react/chunknorrisassesment/src/index.js",[],["11","12"],"/home/petit/projects/react/chunknorrisassesment/src/App.js",["13","14"],{"ruleId":"15","replacedBy":"16"},{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","severity":1,"message":"20","line":29,"column":7,"nodeType":"21","messageId":"22","endLine":29,"endColumn":15},{"ruleId":"19","severity":1,"message":"23","line":33,"column":7,"nodeType":"21","messageId":"22","endLine":33,"endColumn":17},"no-native-reassign",["24"],"no-negated-in-lhs",["25"],"no-unused-vars","'likeJoke' is assigned a value but never used.","Identifier","unusedVar","'unlikeJoke' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
+[{"/home/petit/projects/react/chunknorrisassesment/src/index.js":"1","/home/petit/projects/react/chunknorrisassesment/src/App.js":"2","/home/petit/projects/react/chunknorrisassesment/src/JokeCard.js":"3"},{"size":428,"mtime":1627977527326,"results":"4","hashOfConfig":"5"},{"size":2597,"mtime":1629268175993,"results":"6","hashOfConfig":"5"},{"size":1702,"mtime":1629268525648,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"x0imxm",{"filePath":"11","messages":"12","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"13","messages":"14","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/petit/projects/react/chunknorrisassesment/src/index.js",[],["15","16"],"/home/petit/projects/react/chunknorrisassesment/src/App.js",["17","18","19"],"/home/petit/projects/react/chunknorrisassesment/src/JokeCard.js",["20"],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},{"ruleId":"25","severity":1,"message":"26","line":18,"column":7,"nodeType":"27","messageId":"28","endLine":18,"endColumn":15},{"ruleId":"25","severity":1,"message":"29","line":23,"column":7,"nodeType":"27","messageId":"28","endLine":23,"endColumn":17},{"ruleId":"25","severity":1,"message":"30","line":35,"column":11,"nodeType":"27","messageId":"28","endLine":35,"endColumn":18},{"ruleId":"25","severity":1,"message":"31","line":17,"column":7,"nodeType":"27","messageId":"28","endLine":17,"endColumn":15},"no-native-reassign",["32"],"no-negated-in-lhs",["33"],"no-unused-vars","'likeJoke' is assigned a value but never used.","Identifier","unusedVar","'unlikeJoke' is assigned a value but never used.","'classes' is assigned a value but never used.","'Category' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
